move PendingApplications
